Externalize routes
